fix: moved useDailyPlans into correct folder
