+++ api/useDailyPlans.tsx
@@ -199,125 +199,7 @@ const useDailyPlans = (status?: DailyPlanStatus) => {
     confirmDailyPlanning,
     addTodoToDailyPlan,
     removeTodoFromDailyPlan,
-    // finishDailyTaskList,
-    // finishDailyTask,
   };
 };
 
 export default useDailyPlans;
-
-/**
-const useDailyPlans = (status?: DailyPlanStatus) => {
-  const handleUndoFinishDailyTaskList =
-    (dailyPlan: DailyPlan, toaster: typeof toast) => async () => {
-      const updated: DailyPlan[] = flow(
-        union([dailyPlan]),
-        sortBy((p) => -p.day.getTime())
-      )(dailyPlans);
-      mutate(updated, false);
-      const { data, errors } = await client.models.DailyPlan.update({
-        id: dailyPlan.id,
-        status: "OPEN",
-      });
-      if (errors)
-        handleApiErrors(errors, "Undoing finishing daily plan's task failed");
-      if (updated) mutate(updated);
-      if (!data) return;
-      toaster({
-        title: "Re-opened Daily task list",
-        description: `Daily task list with title “${
-          dailyPlan.dayGoal
-        }” for ${format(dailyPlan.day, "PP")} is open again.`,
-      });
-      return data.id;
-    };
-
-  const finishDailyTaskList = async (
-    dailyPlanId: string,
-    toaster: typeof toast
-  ) => {
-    const oldDailyPlan = dailyPlans?.find((p) => p.id === dailyPlanId);
-    const updated: DailyPlan[] | undefined = dailyPlans?.map((p) =>
-      p.id !== dailyPlanId ? p : { ...p, status: "DONE" }
-    );
-    if (updated) mutate(updated, false);
-    const { data, errors } = await client.models.DailyPlan.update({
-      id: dailyPlanId,
-      status: "DONE",
-    });
-    if (errors) handleApiErrors(errors, "Finishing task list failed");
-    if (updated) mutate(updated);
-    if (!data) return;
-    toaster({
-      title: "Daily task list completed",
-      description: `You completed the daily task list with title “${
-        data.dayGoal
-      }” for ${format(data.day, "PP")}`,
-      action: oldDailyPlan && (
-        <ToastAction
-          altText="Undo finishing task list"
-          onClick={handleUndoFinishDailyTaskList(oldDailyPlan, toaster)}
-        >
-          Undo
-        </ToastAction>
-      ),
-    });
-    return data.id;
-  };
-
-  const finishTask = async (
-    activityId: string,
-    index: number,
-    finished: boolean
-  ) => {
-    const notes = await getActivityNotes(activityId);
-    if (!notes) return;
-    const updated = updateTaskStatus(notes, index, finished);
-    const { hasOpenTasks } = getTasksData(updated);
-    const { data, errors } = await client.models.Activity.update({
-      id: activityId,
-      hasOpenTasks: hasOpenTasks ? "true" : "false",
-      formatVersion: 2,
-      notes: null,
-      notesJson: JSON.stringify(updated),
-    });
-    if (errors) handleApiErrors(errors, "Task status updated");
-    if (!data) return;
-    toast({ title: "Task status updated" });
-    return getTaskByIndex(updated, index);
-  };
-
-  const finishDailyTask = async (
-    dailyPlanId: string,
-    { activityId, index }: DailyPlanTodo,
-    finished: boolean
-  ) => {
-    const updated: DailyPlan[] | undefined = dailyPlans?.map((p) =>
-      p.id !== dailyPlanId
-        ? p
-        : {
-            ...p,
-            tasks: p.tasks.map((t) =>
-              t.activityId !== activityId || t.index !== index
-                ? t
-                : {
-                    ...t,
-                    done: finished,
-                  }
-            ),
-          }
-    );
-    if (updated) mutate(updated, false);
-    const updatedTask = await finishTask(activityId, index, finished);
-    const { data, errors } = await client.models.DailyPlanTask.update({
-      dailyPlanId,
-      activityId: activityId,
-      taskIndex: index,
-      task: JSON.stringify(updatedTask),
-    });
-    if (errors) handleApiErrors(errors, "Updating daily task item failed");
-    if (updated) mutate(updated);
-    return data;
-  };
-};
- */
